+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/util/NativeConfiguration.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.orm.jbt.util;
+package org.hibernate.tool.orm.jbt.internal.util;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -19,8 +19,6 @@ import org.hibernate.cfg.NamingStrategy;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Table;
 import org.hibernate.tool.api.reveng.RevengStrategy;
-import org.hibernate.tool.orm.jbt.internal.util.ExtendedConfiguration;
-import org.hibernate.tool.orm.jbt.internal.util.MetadataHelper;
 import org.w3c.dom.Document;
 import org.xml.sax.EntityResolver;
